[1] Added casts in printfs using %ll type formats to eliminate warnings on x86_64 platform. [1.5] (Will rebuild on ARM64.) [2] Cursory runs of affected utilities, functional tests, unit tests. [2.5] BY NO MEANS COMPLETE. [3] Builds.

+++ Diminuto/fun/adcrheostat.c
@@ -207,7 +207,7 @@ int main(int argc, char ** argv) {
 
             sustain = SUSTAIN;
 
-            printf("%s: PWM %d %% ADC %.3f v Period %lld ms\n", program, duty, volts, elapsed);
+            printf("%s: PWM %d %% ADC %.3f v Period %lld ms\n", program, duty, volts, (long long int)elapsed);
 
             /*
              * Measure allows us to read the voltage more than once.
